Extract interfaces into types.ts and api wrapper for unclaimed pitches
